+++ server/api_tournament.go
@@ -32,6 +32,60 @@ import (
 	"google.golang.org/protobuf/types/known/wrapperspb"
 )
 
+func (s *ApiServer) DeleteTournamentRecord(ctx context.Context, in *api.DeleteTournamentRecordRequest) (*emptypb.Empty, error) {
+	userID := ctx.Value(ctxUserIDKey{}).(uuid.UUID)
+
+	// Before hook.
+	if fn := s.runtime.BeforeDeleteTournamentRecord(); fn != nil {
+		beforeFn := func(clientIP, clientPort string) error {
+			result, err, code := fn(ctx, s.logger, userID.String(), ctx.Value(ctxUsernameKey{}).(string), ctx.Value(ctxVarsKey{}).(map[string]string), ctx.Value(ctxExpiryKey{}).(int64), clientIP, clientPort, in)
+			if err != nil {
+				return status.Error(code, err.Error())
+			}
+			if result == nil {
+				// If result is nil, requested resource is disabled.
+				s.logger.Warn("Intercepted a disabled resource.", zap.Any("resource", ctx.Value(ctxFullMethodKey{}).(string)), zap.String("uid", userID.String()))
+				return status.Error(codes.NotFound, "Requested resource was not found.")
+			}
+			in = result
+			return nil
+		}
+
+		// Execute the before function lambda wrapped in a trace for stats measurement.
+		err := traceApiBefore(ctx, s.logger, s.metrics, ctx.Value(ctxFullMethodKey{}).(string), beforeFn)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	if in.TournamentId == "" {
+		return nil, status.Error(codes.InvalidArgument, "Invalid tournament ID.")
+	}
+
+	if err := TournamentRecordDelete(ctx, s.logger, s.db, s.leaderboardCache, s.leaderboardRankCache, userID, in.TournamentId, userID.String()); err != nil {
+		switch err {
+		case ErrLeaderboardNotFound:
+			return nil, status.Error(codes.NotFound, "Tournament not found.")
+		case ErrLeaderboardAuthoritative:
+			return nil, status.Error(codes.PermissionDenied, "Tournament only allows authoritative score deletions.")
+		default:
+			return nil, status.Error(codes.Internal, "Error deleting score from tournament.")
+		}
+	}
+
+	// After hook.
+	if fn := s.runtime.AfterDeleteTournamentRecord(); fn != nil {
+		afterFn := func(clientIP, clientPort string) error {
+			return fn(ctx, s.logger, userID.String(), ctx.Value(ctxUsernameKey{}).(string), ctx.Value(ctxVarsKey{}).(map[string]string), ctx.Value(ctxExpiryKey{}).(int64), clientIP, clientPort, in)
+		}
+
+		// Execute the after function lambda wrapped in a trace for stats measurement.
+		traceApiAfter(ctx, s.logger, s.metrics, ctx.Value(ctxFullMethodKey{}).(string), afterFn)
+	}
+
+	return &emptypb.Empty{}, nil
+}
+
 func (s *ApiServer) JoinTournament(ctx context.Context, in *api.JoinTournamentRequest) (*emptypb.Empty, error) {
 	userID := ctx.Value(ctxUserIDKey{}).(uuid.UUID)
 	username := ctx.Value(ctxUsernameKey{}).(string)
@@ -301,17 +355,18 @@ func (s *ApiServer) WriteTournamentRecord(ctx context.Context, in *api.WriteTour
 
 	record, err := TournamentRecordWrite(ctx, s.logger, s.db, s.leaderboardCache, s.leaderboardRankCache, userID, in.GetTournamentId(), userID, username, in.GetRecord().GetScore(), in.GetRecord().GetSubscore(), in.GetRecord().GetMetadata(), in.GetRecord().GetOperator())
 	if err != nil {
-		if err == runtime.ErrTournamentMaxSizeReached {
+		switch err {
+		case runtime.ErrTournamentMaxSizeReached:
 			return nil, status.Error(codes.FailedPrecondition, "Tournament has reached max size.")
-		} else if err == runtime.ErrTournamentAuthoritative {
+		case runtime.ErrTournamentAuthoritative:
 			return nil, status.Error(codes.PermissionDenied, "Tournament only allows authoritative score submissions.")
-		} else if err == runtime.ErrTournamentWriteMaxNumScoreReached {
+		case runtime.ErrTournamentWriteMaxNumScoreReached:
 			return nil, status.Error(codes.FailedPrecondition, "Reached allowed max number of score attempts.")
-		} else if err == runtime.ErrTournamentWriteJoinRequired {
+		case runtime.ErrTournamentWriteJoinRequired:
 			return nil, status.Error(codes.FailedPrecondition, "Must join tournament before attempting to write value.")
-		} else if err == runtime.ErrTournamentOutsideDuration {
+		case runtime.ErrTournamentOutsideDuration:
 			return nil, status.Error(codes.FailedPrecondition, "Tournament is not active and cannot accept new scores.")
-		} else {
+		default:
 			return nil, status.Error(codes.Internal, "Error writing score to tournament.")
 		}
 	}